pyproject.toml
[project]
name = "interpolation"
version = "0.1.0"
dynamic = ["version"]
description = "Educational project with interpolation algorithms implementation"
authors = [
    {name = "olyatech", email = "terekhova.oa@phystech.edu"}
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Typing :: Typed",
]
dependencies = ["click (>=8.1.8,<9.0.0)", "pillow (>=11.1.0,<12.0.0)", "numpy (>=2.2.6,<3.0.0)"]

[project.urls]
"Homepage" = "https://github.com/olyatech/Interpolation"

[tool.poetry]
name = "interpolation"

[tool.poetry.scripts]
resize = "interpolation.cli:resize"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.10"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"

[tool.ruff]
select = ["E", "W", "F", "D", "N", "B", "I", "C90", "ANN"]
unfixable = []
ignore = [
    "D203", # ignoring because of incompatibility with other checks
    "D213", # ignoring because of incompatibility with other checks
    "ANN101",  # Types for self/cls
]
line-length = 120
indent-width = 4

[tool.ruff.format]
quote-style = "double"

[tool.ruff.isort]
known-first-party = ["interpolation"]
required-imports = ["from __future__ import annotations"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]